reminderService/requirements.txt
# asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
mysqlclient==2.1.0
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
uwsgi==2.0.17
